requirements.txt
Django
djangorestframework
django-environ
requests
